// src/app/models/response.interfaces.location.ts
export interface ResponseLocation {
    info:    Info;
    results: ResultLocation[];
}

export interface Info {
    count: number;
    pages: number;
    next:  string;
    prev:  null;
}

export interface ResultLocation {
    id:        number;
    name:      string;
    type:      string;
    dimension: string;
    residents: string[];
    url:       string;
    created:   Date;
}